fix: use platform-specific persistence validation queries
