+++ data/persistence/src/jsMain/kotlin/com/wire/kalium/persistence/db/DatabaseValidation.kt
@@ -1,0 +1,60 @@
+/*
+ * Wire
+ * Copyright (C) 2026 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ */
+
+package com.wire.kalium.persistence.db
+
+import app.cash.sqldelight.async.coroutines.await
+import app.cash.sqldelight.async.coroutines.awaitMigrate
+import app.cash.sqldelight.async.coroutines.awaitQuery
+import app.cash.sqldelight.db.QueryResult
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.db.SqlSchema
+import kotlin.coroutines.cancellation.CancellationException
+
+@Suppress("TooGenericExceptionCaught")
+actual suspend fun SqlDriver.migrate(sqlSchema: SqlSchema<QueryResult.AsyncValue<Unit>>): Boolean {
+    val oldVersion = awaitQuery(null, "PRAGMA user_version;", {
+        it.next().await()
+        it.getLong(0)
+    }, 0) ?: return false
+
+    val newVersion = sqlSchema.version
+    return try {
+        if (oldVersion != newVersion) {
+            sqlSchema.awaitMigrate(this, oldVersion, newVersion)
+        }
+        true
+    } catch (exception: CancellationException) {
+        throw exception
+    } catch (_: Exception) {
+        false
+    }
+}
+
+actual suspend fun SqlDriver.checkFKViolations(): Boolean {
+    var result = false
+    awaitQuery(null, "PRAGMA foreign_key_check;", {
+        // foreign_key_check returns rows when the backup is corrupted.
+        if (it.next().await()) {
+            result = true
+        }
+        Unit
+    }, 0, null)
+
+    return result
+}
